Cleanup: move planner/deco related declarations planner/deco.h
