+++ core/deco.c
@@ -21,8 +21,10 @@
 #include <string.h>
 #include <assert.h>
 
+#include "deco.h"
 #include "ssrf.h"
 #include "dive.h"
+#include "gas.h"
 #include "subsurface-string.h"
 #include "errorhelper.h"
 #include "planner.h"
