+++ tests/testplan.cpp
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-2.0
 #include "testplan.h"
+#include "core/deco.h"
 #include "core/dive.h"
 #include "core/planner.h"
 #include "core/qthelper.h"
